Bulk upload feature for Gross Domestic Product

+++ backend/database/migrations/2022_12_01_125815_create_passenger_road_transport_data_table.php
@@ -17,12 +17,12 @@ class CreatePassengerRoadTransportDataTable extends Migration
         Schema::create('passenger_road_transport_data', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('year')->unique();
-            $table->bigInteger('number_of_passengers_carried')->nullable();
-            $table->bigInteger('number_of_vehicles_in_fleet')->nullable();
-            $table->bigInteger('revenue_from_operation')->nullable();
-            $table->bigInteger('number_of_employees')->nullable();
-            $table->bigInteger('annual_cost_of_vehicle_maintenance')->nullable();
-            $table->bigInteger('number_of_accidents')->nullable();
+            $table->float('number_of_passengers_carried')->nullable();
+            $table->float('number_of_vehicles_in_fleet')->nullable();
+            $table->float('revenue_from_operation')->nullable();
+            $table->float('number_of_employees')->nullable();
+            $table->float('annual_cost_of_vehicle_maintenance')->nullable();
+            $table->float('number_of_accidents')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
