+++ backend/database/migrations/2022_12_05_141425_create_railway_passengers_table.php
@@ -17,15 +17,15 @@ class CreateRailwayPassengersTable extends Migration
         Schema::create('railway_passengers', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('year')->unique();
-            $table->bigInteger('number_of_urban_passengers_carried')->nullable();
-            $table->bigInteger('number_of_regional_passengers_carried')->nullable();
-            $table->bigInteger('freight_carried')->nullable();
-            $table->bigInteger('number_of_freight_trains')->nullable();
-            $table->bigInteger('number_of_passenger_trains')->nullable();
-            $table->bigInteger('freight_revenue_generation')->nullable();
-            $table->bigInteger('passenger_revenue_generation')->nullable();
-            $table->bigInteger('passenger_fuel_consumption_rate')->nullable();
-            $table->bigInteger('freight_fuel_consumption_rate')->nullable();
+            $table->float('number_of_urban_passengers_carried')->nullable();
+            $table->float('number_of_regional_passengers_carried')->nullable();
+            $table->float('freight_carried')->nullable();
+            $table->float('number_of_freight_trains')->nullable();
+            $table->float('number_of_passenger_trains')->nullable();
+            $table->float('freight_revenue_generation')->nullable();
+            $table->float('passenger_revenue_generation')->nullable();
+            $table->float('passenger_fuel_consumption_rate')->nullable();
+            $table->float('freight_fuel_consumption_rate')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
