+++ backend/database/migrations/2024_03_18_104854_create_gross_domestic_products_table.php
@@ -17,13 +17,13 @@ class CreateGrossDomesticProductsTable extends Migration
         Schema::create('gross_domestic_products', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('year');
-            $table->integer('transportation_and_storage')->nullable();
-            $table->integer('road_transport')->nullable();
-            $table->integer('rail_transport_and_pipelines')->nullable();
-            $table->integer('water_transport')->nullable();
-            $table->integer('air_transport')->nullable();
-            $table->integer('transport_services')->nullable();
-            $table->integer('post_and_courier_services')->nullable();
+            $table->float('transportation_and_storage')->nullable();
+            $table->float('road_transport')->nullable();
+            $table->float('rail_transport_and_pipelines')->nullable();
+            $table->float('water_transport')->nullable();
+            $table->float('air_transport')->nullable();
+            $table->float('transport_services')->nullable();
+            $table->float('post_and_courier_services')->nullable();
             $table->timestamps();
         });
     }
