+++ backend/database/migrations/2024_03_18_123059_create_vehicle_plate_productions_table.php
@@ -18,7 +18,7 @@ class CreateVehiclePlateProductionsTable extends Migration
             $table->increments('id');
             $table->integer('vehicle_category');
             $table->integer('year');
-            $table->bigInteger('value')->nullable();
+            $table->float('value')->nullable();
             $table->timestamps();
         });
     }
