+++ backend/database/migrations/2024_03_18_132420_create_license_renewals_table.php
@@ -17,15 +17,15 @@ class CreateLicenseRenewalsTable extends Migration
         Schema::create('license_renewals', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('year');
-            $table->bigInteger('car')->nullable();
-            $table->bigInteger('van')->nullable();
-            $table->bigInteger('lorry')->nullable();
-            $table->bigInteger('mini_bus')->nullable();
-            $table->bigInteger('big_bus')->nullable();
-            $table->bigInteger('tanker')->nullable();
-            $table->bigInteger('trailer')->nullable();
-            $table->bigInteger('tipper')->nullable();
-            $table->bigInteger('tractor')->nullable();
+            $table->float('car')->nullable();
+            $table->float('van')->nullable();
+            $table->float('lorry')->nullable();
+            $table->float('mini_bus')->nullable();
+            $table->float('big_bus')->nullable();
+            $table->float('tanker')->nullable();
+            $table->float('trailer')->nullable();
+            $table->float('tipper')->nullable();
+            $table->float('tractor')->nullable();
             $table->text('state');
             $table->timestamps();
         });
