+++ backend/database/migrations/2024_03_18_141114_create_fleet_accidents_table.php
@@ -18,8 +18,8 @@ class CreateFleetAccidentsTable extends Migration
             $table->increments('id');
             $table->integer('year');
             $table->text('transport_company')->nullable();
-            $table->bigInteger('vehicle')->nullable();
-            $table->bigInteger('number_of_accidents')->nullable();
+            $table->float('vehicle')->nullable();
+            $table->float('number_of_accidents')->nullable();
             $table->timestamps();
         });
     }
