+++ backend/database/migrations/2024_03_19_111817_create_air_mode_datas_table.php
@@ -18,14 +18,14 @@ class CreateAirModeDatasTable extends Migration
             $table->increments('id');
             $table->string('airport');
             $table->integer('year');
-            $table->bigInteger('domestic_passengers_traffic')->nullable();
-            $table->bigInteger('international_passengers_traffic')->nullable();
-            $table->bigInteger('aircraft_traffic_domestic')->nullable();
-            $table->bigInteger('aircraft_traffic_international')->nullable();
-            $table->bigInteger('cargo_traffic_domestic')->nullable();
-            $table->bigInteger('cargo_traffic_international')->nullable();
-            $table->bigInteger('mail_traffic_domestic')->nullable();
-            $table->bigInteger('mail_traffic_international')->nullable();
+            $table->float('domestic_passengers_traffic')->nullable();
+            $table->float('international_passengers_traffic')->nullable();
+            $table->float('aircraft_traffic_domestic')->nullable();
+            $table->float('aircraft_traffic_international')->nullable();
+            $table->float('cargo_traffic_domestic')->nullable();
+            $table->float('cargo_traffic_international')->nullable();
+            $table->float('mail_traffic_domestic')->nullable();
+            $table->float('mail_traffic_international')->nullable();
             $table->timestamps();
         });
     }
